requirements/local.txt
-r ./base.txt

django-nplusone==0.0.5
django-debug-toolbar==3.2.2
djhtml==1.4.9
flake8==4.0.1
black==21.10b0
factory-boy==3.2.1
django-extensions==3.1.3
django-coverage-plugin==2.0.1
pre-commit==2.15.0
pygraphviz==1.7
pydot==1.4.2
coverage==6.1.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-randomly==3.10.1
pytest-xdist==2.4.0
pyparsing==3.0.4
pretty-errors==1.2.24
tryceratops==0.6.0